pyproject.toml
[project]
name = "matterworks-lipid"
version = "0.1.0"
description = "Lipid identification from MS2 spectra via graph neural networks"
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "torch>=2.2.0",
    "numpy>=1.26.0,<2",
    "pandas>=2.1.0",
    "faiss-cpu>=1.7.4",
    "pyarrow>=23.0.1",
    "fastparquet>=2025.12.0",
    "matplotlib>=3.10.0",
    "seaborn>=0.13.2",
    "nbconvert>=7.17.0",
    "ipykernel>=7.2.0",
    "pytest>=8.0.0",
    "matchms>=0.28.0,<0.31",
    # DreaMS transitive deps (installed editable from ./DreaMS)
    "h5py>=3.11",
    "pytorch-lightning>=2.0",
    "umap-learn>=0.5",
    "igraph>=0.11",
    "plotly>=5.20",
    "ase>=3.22",
    "fire>=0.6",
    "pandarallel>=1.6",
    "huggingface-hub>=0.33",
    "pyopenms>=3.4",
    "wandb>=0.16",
    "rdkit",
    "transformers>=4.30",
    "tqdm>=4.65",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
markers = ["slow: marks tests that load large checkpoints or models (deselect with '-m \"not slow\"')"]
